use new oq config deom model; revert to float64 values;

+++ toshi_hazard_store/model/revision_4/extract_classical_hdf5.py
@@ -147,7 +147,7 @@ def rlzs_to_record_batch_reader(
             yield batch
 
     # create a schema...
-    values_type = pa.list_(pa.float32())  # TODO CHECK if this is enough res, or float32 float64
+    values_type = pa.list_(pa.float64())  # TODO CHECK if this is enough res, or float32 float64
     vs30_type = pa.int32()
     dict_type = pa.dictionary(pa.int32(), pa.string(), True)
     schema = pa.schema(
